+++ CHANGELOG.md
@@ -6,6 +6,8 @@ unsoundness!
 * Add common `Builder` trait to `EntityBuilder` and `LazyBuilder` ([#426])
 * Add common `MarkedBuilder` trait to `EntityBuilder` and `LazyBuilder` ([#426])
 
+[#426]: https://github.com/slide-rs/specs/pull/426
+
 # 0.11.4
 
 * Add `build_entity()` to `EntitiesRes` so you can use builder syntax in a system.